refactor: Remove dao package
