+++ src/main/java/kitchenpos/menu/domain/MenuGroupRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.menu.dao;
+package kitchenpos.menu.domain;
 
 import kitchenpos.menu.domain.MenuGroup;
