+++ src/main/java/kitchenpos/order/domain/Order.java
@@ -1,8 +1,5 @@
 package kitchenpos.order.domain;
 
-import kitchenpos.order.dao.OrderStatus;
-import kitchenpos.order.dao.OrderType;
-
 import javax.persistence.*;
 import java.time.LocalDateTime;
 import java.util.List;
