+++ src/main/java/kitchenpos/order/domain/OrderStatus.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.dao;
+package kitchenpos.order.domain;
 
 public enum OrderStatus {
     WAITING, ACCEPTED, SERVED, DELIVERING, DELIVERED, COMPLETED
