+++ src/main/java/kitchenpos/order/domain/OrderType.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.dao;
+package kitchenpos.order.domain;
 
 public enum OrderType {
     DELIVERY, TAKEOUT, EAT_IN
