+++ src/test/java/kitchenpos/application/InMemoryMenuGroupRepository.java
@@ -1,7 +1,7 @@
 package kitchenpos.application;
 
 import kitchenpos.menu.domain.MenuGroup;
-import kitchenpos.menu.dao.MenuGroupRepository;
+import kitchenpos.menu.domain.MenuGroupRepository;
 
 import java.util.*;
