+++ src/test/java/kitchenpos/application/ProductServiceTest.java
@@ -1,9 +1,9 @@
 package kitchenpos.application;
 
 import kitchenpos.menu.domain.Menu;
-import kitchenpos.menu.dao.MenuRepository;
+import kitchenpos.menu.domain.MenuRepository;
 import kitchenpos.product.domain.Product;
-import kitchenpos.product.dao.ProductRepository;
+import kitchenpos.product.domain.ProductRepository;
 import kitchenpos.infra.PurgomalumClient;
 import kitchenpos.product.application.ProductService;
 import org.junit.jupiter.api.BeforeEach;
